#12 Detach component from CourseDetail
