+++ src/features/course/components/CourseDetail/CourseDetail.jsx
@@ -1,51 +1,23 @@
-// import Classname:
-import classNames from 'classnames';
-// import hook:
-import { useState } from 'react';
 // import Component:
 import Tag from '../../../../components/ui/Tag';
 import SubNav from '../SubNav';
-import Chapter from './Chapter';
+import AchieveList from './AchieveList';
+import CourseContent from './CourseContent';
 import ModalRegisterCourse from './ModalRegisterCourse';
+import CourseInfo from './CourseInfo';
+import QualityItem from './QualityItem';
 // import Icon:
 import { AiFillStar, AiOutlineTeam } from 'react-icons/ai';
 import { BiBookmarks, BiBookOpen } from 'react-icons/bi';
-import { TiTick } from 'react-icons/ti';
 import { FaRegLightbulb } from 'react-icons/fa';
 // import Image:
 import bannerImg from '../../../../assets/images/courseDetailBanner.png';
 // import Data:
-import { courseDetail, chapters } from '../data';
+import { courseDetail } from '../data';
 // import hard data:
 import { qualityUs } from '../../../../data/constants';
 
-// Declare TickItem component:
-const TickItem = ({ children }) => (
-  <li className="mt-2 flex gap-3 items-center">
-    <div className="w-[24px] h-[24px] bg-ac_green flex flex-shrink-0 justify-center rounded-sm items-center">
-      <TiTick className="text-white" />
-    </div>
-    <p>{children}</p>
-  </li>
-);
-
-// Declare QualityItem component:
-const QualityItem = ({ children, icon }) => (
-  <div className="border-2 border-primary rounded-lg pb-3 pt-6 px-5 relative flex items-center">
-    <p>{children}</p>
-    <div className="absolute top-0 -translate-y-1/2 left-1/2 -translate-x-1/2 bg-white">
-      <div className="relative w-[40px] h-[40px] before:block before:w-[40px] before:h-[40px] before:rounded-full before:bg-primary before:opacity-50">
-        <div className="text-primary text-[24px] absolute bottom-0 -translate-y-1/3  right-0 -translate-x-1/3">
-          {icon}
-        </div>
-      </div>
-    </div>
-  </div>
-);
-
 function CourseDetail() {
-  const [openAll, setOpenAll] = useState(false);
-
   // Fill color for Star rating icon:
   const fillStar = (avgRate, num) => {
     return Math.round(avgRate) < num ? (
@@ -105,140 +77,30 @@ function CourseDetail() {
         </div>
       </div>
 
+      {/* Main content Page */}
       <div className="mx-6  md:mx-16 lg:flex lg:flex-row-reverse lg:gap-5 lg:mx-[100px]">
-        {/* ModalRegisterCourse component */}
         <div className="lg:w-4/12 lg:relative">
+          {/* ModalRegisterCourse component */}
           <div className="mt-10 min-h-[400px] md:w-1/2 md:mx-auto lg:w-full lg:sticky top-5 lg:mt-[-400px] xl:mt-[-500px]">
             <ModalRegisterCourse course={courseDetail} />
           </div>
         </div>
 
-        {/* Main content CourseDetail Page */}
+        {/* Content CourseDetail Page */}
         <div className="lg:w-8/12">
+          {/* Achieve list component */}
           <div className="mt-10 md:shadow-lg md:p-4 md:rounded-lg lg:p-8">
-            {/* Achieve after course */}
-            <h3 className="text-body-lg text-center font-medium lg:text-left">Bạn nhận được gì sau khóa học?</h3>
-            <div className="grid gap-3 mt-4 md:grid-cols-2 lg:gap-4 lg:mt-8">
-              {/* List achieve */}
-              {courseDetail?.achieves?.map((achieve, index) => (
-                <div key={achieve.id} className="flex items-center gap-4">
-                  <div className="w-[30px] h-[30px] bg-primary flex flex-shrink-0 justify-center rounded-md items-center">
-                    <span className="text-white">{index + 1}</span>
-                  </div>
-                  <p>{achieve.content}</p>
-                </div>
-              ))}
-            </div>
+            <AchieveList achieveList={courseDetail?.achieves} />
           </div>
 
-          {/* Course detail infomation */}
+          {/* Course infomation component */}
           <div className="mt-10 md:bg-bg_light_gray md:p-4 md:rounded-lg lg:p-8">
-            <h3 className="text-body-lg text-center font-medium mb-4 lg:text-left lg:mb-8">Thông tin khóa học</h3>
-            {/* Author info */}
-            <div>
-              <p className="font-medium">
-                <i>Bài học được biên soạn và giảng dạy bởi nhóm tác giả:</i>
-              </p>
-              {/* List authors */}
-              <ul className="list-disc pl-5 text-body-md font-normal mt-3">
-                {courseDetail?.authors?.map((author) => (
-                  <li key={author.id} className="mt-2">
-                    {author.content}
-                  </li>
-                ))}
-              </ul>
-            </div>
-
-            {/* Target Users info */}
-            <div className="mt-8">
-              <p className="font-medium">
-                <i>Nhóm đối tượng tham gia khóa học:</i>
-              </p>
-              {/* List target users */}
-              <ul className="text-body-md font-normal mt-3 grid gap-3">
-                {courseDetail?.targetUsers?.map((target) => (
-                  <TickItem key={target.id}>{target.content}</TickItem>
-                ))}
-              </ul>
-            </div>
-
-            {/* Skill info */}
-            <div className="mt-8">
-              <p className="font-medium">
-                <i>Khóa học bổ trợ các kỹ năng:</i>
-              </p>
-              {/* List skill */}
-              <ul className="text-body-md font-normal mt-3 grid gap-3">
-                {courseDetail?.skills?.map((skill) => (
-                  <TickItem key={skill.id}>{skill.content}</TickItem>
-                ))}
-              </ul>
-            </div>
-
-            {/* Other info */}
-            <div className="mt-8">
-              <p className="font-medium">
-                <i>Tổng quan:</i>
-              </p>
-              <ul className="list-disc pl-5 text-body-md font-normal mt-3">
-                {/* Level course info */}
-                <li className="mt-2">
-                  Mức độ:
-                  <b
-                    // Style for type of level
-                    className={classNames({
-                      'text-ac_blue': courseDetail.level === 'basic',
-                      'text-ac_orange': courseDetail.level === 'general',
-                      'text-ac_red': courseDetail.level === 'advance',
-                    })}
-                  >
-                    {/* Check level */}
-                    {courseDetail.level === 'basic' && 'cơ bản'}
-                    {courseDetail.level === 'general' && 'phổ thông'}
-                    {courseDetail.level === 'advance' && 'nâng cao'}
-                  </b>
-                </li>
-
-                {/* List  other info  */}
-                {courseDetail?.other?.map((info) => (
-                  <li key={info.id} className="mt-2">
-                    {info.content}
-                  </li>
-                ))}
-              </ul>
-            </div>
+            <CourseInfo course={courseDetail} />
           </div>
 
-          {/* Couse Content */}
+          {/* Couse Content component*/}
           <div className="mt-10 md:mt-20">
-            <h3 className="text-body-lg text-center font-medium mb-4 md:mb-8 lg:text-left">Nội dung khóa học</h3>
-            <div className="md:flex md:justify-between mb-2 md:mb-4">
-              <p className="text-body-sm">
-                <b>{courseDetail.totalChapter}</b> chương .{' '}
-                <span className="hidden md:inline-block">
-                  <b>{courseDetail.totalLesson}</b> bài học .{' '}
-                </span>
-                thời lượng{' '}
-                <b>
-                  {courseDetail.totalVideoTime.hour} giờ {courseDetail.totalVideoTime.minutes} phút
-                </b>
-              </p>
-
-              {/* Button expand course content */}
-              <span
-                className="text-primary text-body-sm font-semibold cursor-pointer"
-                onClick={() => setOpenAll(!openAll)}
-              >
-                {openAll ? 'Thu gọn' : 'Mở rộng tất cả'}
-              </span>
-            </div>
-
-            {/* List chapter of the course */}
-            <div className="grid gap-2">
-              {chapters.map((chapter) => (
-                <Chapter key={chapter.id} chapter={chapter} openAll={openAll} />
-              ))}
-            </div>
+            <CourseContent course={courseDetail} />
           </div>
         </div>
       </div>
